Move game-board DOM to ui.js file
